refactor: TitleStackView 분리
